# lesson-booking/openapi.yaml
openapi: 3.1.0
info:
  title: Lesson Booking API
  version: 0.1.0
  description: |
    レッスン予約サービス。複数 Google カレンダー連携で空き枠を出し、40 日先まで予約できる。
    レッスン開始まで 14 日未満のキャンセル・変更はメッセージ+対応方法(3 択)を添えた変更要求となり、主催者の承認で確定する。
    詳細は docs/spec.md。
servers:
  - url: http://localhost:8787
security:
  - bearerAuth: []
  - devHeader: []
...
paths:
...
  /me:
    get:
      tags: [student, host]
      summary: ログイン中ユーザー
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  email: { type: string }
                  name: { type: string }
                  role: { type: string, enum: [host, student] }
                  host: { oneOf: [{ $ref: '#/components/schemas/Host' }, { type: 'null' }] }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      tags: [student, host]
      summary: 退会(アカウントと全データの削除)
      description: |
        確認文字列が必要(主催者は URL 名、生徒はメールアドレス)。
        主催者: 今後の予約をすべて休講扱いで取り消して生徒に通知 → 有料契約を即時解約 → Google 連携を失効 → 全データ削除。
        生徒: 今後の確定予約が残っていると 409 invalid_state。
        最後にログインアカウント(Supabase Auth)を削除する。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedHost: { type: boolean }
                  deletedStudent: { type: boolean }
                  cancelledBookings: { type: integer }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
  /internal/cron/reminders:
    post:
      tags: [public]
      security: []
      summary: 前日リマインドの送信と、カレンダーへの反映に失敗した予約の再試行(外部 cron から毎時)
      parameters:
        - { name: x-cron-secret, in: header, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  checked: { type: integer }
                  sent: { type: integer }
                  failed: { type: integer }
                  calendarSync:
                    type: object
                    properties:
                      checked: { type: integer }
                      fixed: { type: integer }
        '403': { $ref: '#/components/responses/Error' }
...
  /billing/connect-webhook:
    post:
      tags: [public]
      security: []
      summary: Stripe Connect Webhook(キャンセルフィー決済完了・連結アカウント状態)
      responses:
        '200': { description: OK }
        '403': { $ref: '#/components/responses/Error' }
...
  /billing/webhook:
    post:
      tags: [public]
      security: []
      summary: Stripe Webhook(生ボディ・Stripe-Signature ヘッダで署名検証)
      responses:
        '200': { description: OK }
        '403': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase Auth の access_token }
    devHeader: { type: apiKey, in: header, name: x-dev-user-email, description: AUTH_MODE=dev のときのみ }
...
  responses:
    Error:
      description: エラー
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    enum: [validation, outside_booking_window, late_change_requires_request, forbidden, not_found, slot_unavailable, change_request_pending, invalid_state, plan_limit, calendar_error, internal]
                  message: { type: string }
                  details: {}
  schemas:
    FeeMethod:
      type: string
      enum: [card, bank_transfer, in_person]
      description: クレジットカード / 銀行振込 / 次回レッスン時に手渡し
...
    Host:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string }
        slug: { type: string }
        bio: { type: string }
        plan: { type: string, enum: [free, pro] }
        subscriptionStatus: { type: string, enum: [none, active, past_due, canceled] }
        cancellationFeeAmount: { type: [integer, 'null'] }
        feeMethods: { type: array, items: { $ref: '#/components/schemas/FeeMethod' } }
        bankTransferInfo: { type: string }
        stripeConnectAccountId: { type: [string, 'null'] }
        connectChargesEnabled: { type: boolean }
        organizationId: { type: [string, 'null'] }
        orgPlanActive: { type: boolean, description: 所属教室の契約が有効か(有効ならプロ相当) }
        timezone: { type: string }
        lessonMinutes: { type: integer }
        minLeadMinutes: { type: integer }
        createdAt: { type: string, format: date-time }
...